long

+++ core/src/main/scala/ai/lum/odinson/metadata/Ast.scala
@@ -14,7 +14,7 @@ object Ast {
     case class NotEqual(lhs: Value, rhs: Value) extends BoolExpression
 
     sealed trait Value
-    case class NumberValue(n: Int) extends Value
+    case class NumberValue(n: Long) extends Value
     case class StringValue(s: String) extends Value
     case class FunCall(name: String, args: Seq[Value]) extends Value
